Refactor site building to not take so damn long
