+++ .gitignore
@@ -22,6 +22,7 @@ _site/sources.md
 _site/latest.md
 _site/performance.md
 _site/robotstxt.md
+_site/openai-gptbot-robotstxt.md
 extracts/csv/us-right-wing-hyperlinks-sample.csv
 
 # Byte-compiled / optimized / DLL files
